Dev to Master

+++ src/Controller/AdminActivityController.php
@@ -11,7 +11,7 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
 /**
- * @Route("/superadmin/activity")
+ * @Route("/superadmin/activite")
  */
 class AdminActivityController extends AbstractController
 {
@@ -26,7 +26,7 @@ class AdminActivityController extends AbstractController
     }
 
     /**
-     * @Route("/new", name="admin_activity_new", methods={"GET","POST"})
+     * @Route("/nouvelle", name="admin_activity_new", methods={"GET","POST"})
      */
     public function new(Request $request): Response
     {
@@ -59,7 +59,7 @@ class AdminActivityController extends AbstractController
     }
 
     /**
-     * @Route("/{id}/edit", name="admin_activity_edit", methods={"GET","POST"})
+     * @Route("/{id}/editer", name="admin_activity_edit", methods={"GET","POST"})
      */
     public function edit(Request $request, Activity $activity): Response
     {
@@ -83,12 +83,19 @@ class AdminActivityController extends AbstractController
      */
     public function delete(Request $request, Activity $activity): Response
     {
-        if ($this->isCsrfTokenValid('delete' . $activity->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($activity);
-            $entityManager->flush();
+        $isClientEmpty = $activity->getClients()->isEmpty();
+        $isCoachEmpty = $activity->getCoaches()->isEmpty();
+        $isSpaceEmpty = $activity->getTrainingSpaces()->isEmpty();
+        if ($isClientEmpty &&  $isCoachEmpty && $isSpaceEmpty) {
+            if ($this->isCsrfTokenValid('delete' . $activity->getId(), $request->request->get('_token'))) {
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->remove($activity);
+                $entityManager->flush();
+                $this->addFlash('success', 'L\'activité a été supprimée avec succès.');
+            }
+        } else {
+            $this->addFlash('warning', 'Vous ne pouvez pas supprimer cette activité.');
         }
-
         return $this->redirectToRoute('admin_activity_index');
     }
 }
